# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bookit API Documentation",
        "title": "Bookit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/houses/liked": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Get houses liked by current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.HouseLikeItem"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/houses/{id}/like": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Check if user liked a house",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.HouseLikeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Like a house",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.HouseLikeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Houses"
                ],
                "summary": "Unlike a house",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "House ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.HouseLikeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                }
            }
        },
...
        "schemas.HouseLikeItem": {
            "type": "object",
            "properties": {
                "address_en": {
                    "type": "string",
                    "example": "123 Beach Rd"
                },
                "address_kz": {
                    "type": "string"
                },
                "address_ru": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "liked_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-03-30T12:00:00Z"
                },
                "name_en": {
                    "type": "string",
                    "example": "Beach House"
                },
                "name_kz": {
                    "type": "string",
                    "example": "Жағажай үйі"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Пляжный дом"
                },
                "price": {
                    "type": "integer",
                    "example": 50000
                },
                "slug": {
                    "type": "string",
                    "example": "beach-house"
                }
            }
        },
        "schemas.HouseLikeResponse": {
            "type": "object",
            "properties": {
                "like_count": {
                    "type": "integer",
                    "example": 5
                },
                "liked": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}